requirements.txt
asgiref==3.3.4
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
Django==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jieba3k==0.35.1
jmespath==0.10.0
joblib==1.0.1
lxml==4.6.3
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.3
parsel==1.6.0
Pillow==8.2.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
queuelib==1.6.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
Scrapy==2.5.0
selenium==3.141.0
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.61.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urllib3==1.26.4
w3lib==1.22.0
zope.interface==5.4.0
